infrastructure/requirements.txt
aws_cdk.core==1.94.1
aws-cdk.aws-amplify==1.94.1
aws-cdk.aws-dynamodb==1.94.1
aws-cdk.aws-events-targets==1.94.1
aws-cdk.aws-lambda-event-sources==1.94.1
aws-cdk.aws-lambda==1.94.1
aws-cdk.aws-route53==1.94.1
aws-cdk.aws-s3==1.94.1
aws_cdk.aws_apigatewayv2==1.94.1
aws_cdk.aws_apigatewayv2_integrations==1.94.1
aws_cdk.aws_sqs==1.94.1
aws_cdk.aws_logs==1.94.1
aws_cdk.aws_iam==1.94.1
